async_runtime: Add ability to properly shut down and retrieve result
This adds the ability to AsyncRuntime and ExecutionEngine to properly
shut them down and to retrieve the result of the primary task (Future).
The primary task can be blocked_on or run async in an ExecutionEngine
and the result retrieved or waited on the task finishing later.
Shutdown of engines is also possible.